Time: 113 ms (61.02%), Space: 69 MB (26.27%) - LeetHub

+++ 1482-minimum-number-of-days-to-make-m-bouquets/1482-minimum-number-of-days-to-make-m-bouquets.cpp
@@ -16,7 +16,7 @@ public:
         return b>=m;
     }
     int minDays(vector<int>& bloomDay, int m, int k){
-        long long val = m * 1ll * k * 1ll;
+        long int val = m * 1ll * k * 1ll;
         int n = bloomDay.size(); 
         if (val > n){
             return -1;
